+++ CHANGELOG.md
@@ -10,3 +10,4 @@
 - Add a default `Host` tab that displays information about the docker host
 - Add data usage tab with containers, images and build cache data usage stats
 - Data is now preloaded at startup so that it is instantly available
+- Adjust the API version of the client so that it works with older versions of Docker